+++ src/AGame.cpp
@@ -10,7 +10,7 @@
 
 #include "AGame.hpp"
 
-AGame::AGame(std::string const &name, Vector2<int> size) : name(name), size(size)
+AGame::AGame(std::string const &name, Vector2<double> size) : name(name), size(size)
 {
 
 }
@@ -20,7 +20,7 @@ AGame::~AGame()
 
 }
 
-Vector2<int> const           &AGame::getSize() const
+Vector2<double> const           &AGame::getSize() const
 {
     return (this->size);
 }
